% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGindex.R
\name{HGindex}
\alias{HGindex}
\title{Height growth rate index estimation}
\usage{
HGindex(df, ForT, H0, H1, Hr, Dt)
}
\arguments{
\item{df}{dataframe with input values}

\item{ForT}{Forest type ID}

\item{H0}{height at time 0}

\item{H1}{height at time 1}

\item{Hr}{Height reference}

\item{Dt}{period length (years)}
}
\value{
height growth rate index
}
\description{
Estimate height growth index based on the height increment and at height reference
}
\examples{
HGindex(df=dat2,ForT=dat2$ForType,H0=dat2$H0,H1=dat2$H1,Hr=15,Dt=7)
}
