% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hgrowth.R
\name{Hgrowth}
\alias{Hgrowth}
\title{Height growth estimation}
\usage{
Hgrowth(df, ForT, H0, Dt)
}
\arguments{
\item{df}{dataframe with input values}

\item{ForT}{Forest type ID}

\item{H0}{height at time 0}

\item{Dt}{period length (years)}
}
\value{
Height at time 1 H1
}
\description{
Estimate the height at time 1 based on initial height and period length. See manuscript describing the modeling approach
}
\examples{
Hgrowth(sampledat,ForT=sampledat$ForType,H0=sampledat$H0,Dt=7)
}
