% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGindex.R
\name{HGSIndex}
\alias{HGSIndex}
\title{Height growth rate index estimation}
\usage{
HGSIndex(H0, H1, ForT, Hr, Dt)
}
\arguments{
\item{H0}{height at time 0}

\item{H1}{height at time 1}

\item{ForT}{Forest type identification codes.}

\item{Hr}{Height reference}

\item{Dt}{period length (years)}

\item{df}{dataframe with input values}
}
\value{
height growth rate index
}
\description{
Estimate height growth index based on the height increment and at height reference
}
\examples{
HGSIndex(H0=10, H1=15, ForT='PJ',Hr=15,Dt=7)

}
