% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sampledat}
\alias{sampledat}
\title{Sample data for estimations}
\format{
A \code{tibble} with 3 columns:
\describe{
\item{ForType}{Forest types ID}
\item{H0}{height at time 0}
\item{ Dt}{period length (years)}
}
}
\usage{
sampledat
}
\description{
A sample \code{tibble} that is for siple testing and runnig of examples
}
\keyword{datasets}
